+++ src/main/java/com/binance/chuyennd/ai_ml/hpo/HPOEvaluator.java
@@ -6,11 +6,11 @@ import com.binance.chuyennd.utils.Utils;
 import java.util.Collection;
 
 public class HPOEvaluator {
-    public static double evaluateProfitVelocity(SimulatorMarketLevelTicker1MStopLoss simulator) {
+    public static float evaluateProfitVelocity(SimulatorMarketLevelTicker1MStopLoss simulator) {
         Collection<OrderTargetInfoTest> orders = simulator.allOrderDone.values();
-        if (orders.isEmpty()) return -1000.0; // Phạt nếu không có lệnh nào
+        if (orders.isEmpty()) return -1000.0f; // Phạt nếu không có lệnh nào
 
-        double totalProfit = 0;
+        float totalProfit = 0;
         long totalHoldingTime = 0;
 
         for (OrderTargetInfoTest order : orders) {
@@ -18,7 +18,7 @@ public class HPOEvaluator {
             totalHoldingTime += (order.timeUpdate - order.timeStart);
         }
 
-        double avgHoldingTimeHours = (double) totalHoldingTime / orders.size() / Utils.TIME_HOUR;
+        float avgHoldingTimeHours = (float) totalHoldingTime / orders.size() / Utils.TIME_HOUR;
         if (avgHoldingTimeHours <= 0) avgHoldingTimeHours = 1;
 
         if (totalProfit > 0) {
